Accept luajit as DSO
Some Linux distributions (at least Fedora and RHEL) do not provide
libluajit as an archive and instead only as a DSO.  Using the DSO
does not have any negative impact but the cmake script does not
check for the DSO files and instead fails to find luajit.

This simple change causes cmake to find the file as named in the
above distributions.  Maybe more change or a generalized method are
needed in future.  The prefered way to detect the DSO for Fedora/RHEL
would be to use pkgconfig but I don't propose to use it at this time.