Fix computation of LUAJIT_INCLUDE_DIR
I don't think the current way cmake uses to locate the correct lua.h
works.  Entries like "include/foo.h" cause cmake to look for a file
like "/usr/include/include/foo.h".  I.e., the include/ prefix for
all the entries is wrong as is adding "include" at the end.

This patch removes the prefix.  With it bcc finally compiles on
Fedora.

If this change causes a problem there is something seriously wrong
with cmake.