add MEMBER_(ADDR|READ) helper macros
These helper macros are handy when bcc is not yet able to expand
'dst = src->member;' constructs automatically. With these macro, such
constructs can be written as 'MEMBER_READ(&dst, src, member)' which
preserve readability and functionality. It is also generic enough to fit
a large number of use cases.