Fix freezing of 'test_brb' and 'test_brb2' if netperf is not found
If netperf is not installed or installed at a location that is not
in PATH as recognized by Python, then 'test_brb' and 'test_brb2'
freeze after an OSError is raised. To avoid this, we proactively
check if the 'iperf', 'netserver' and 'netperf' binaries are
available before making the corresponding NSPopen() calls.

Signed-off-by: Sandipan Das <sandipan@linux.vnet.ibm.com>