histogram: support sorting bucket before dumping table
When there are multiple fields in the secondary key of a histogram
(e.g. output the bio latency for each combination of disk and blk cgroup),
if the buckets are not sorted before the iteration, the output order
of tables often does not meet our expectations.

So provide a way to sort them.