bpfd: Add BPFd (BPF daemon)
This patch adds BPFd, a standalone executable designed to provide BCC the
ability to work across system and architecture boundaries.

This is done by loading the BPFd executable onto a remote target device to
have it act as a proxy for whenever a BCC tool wishes to perform an operation on
the system (e.g. load BPF programs, read /proc/kallsyms, attach kprobes, etc.).

This arrangement allows developers to have kernel sources and the LLVM stack
on a separate host machine (e.g. the development machine) instead of needing
to set all these up on the target device, thereby allowing for the drastic
reduction of space required on a target for BCC tools to run. The reduction of
the space requirement, in particular, becomes a much more critical factor
for devices that have more limited disk space (e.g. embedded devices).

In addition, the above set-up also allows developers to run clang on a
different architecture than the target's architecture, thus facilitating
cross-compilation development.

However, the natural disadvantage for cross-developers is that there is a
need to have a copy of the target's kernel sources on the host for the
above set-up to work.

For more information, please check out the README in the original
BPFd repository (https://github.com/joelagnel/bpfd) and this LWN article
explaining the purpose of and how BPFd works in more detail
(https://lwn.net/Articles/744522/)

Signed-off-by: Jazel Canseco <jcanseco@google.com>
Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.com>