locktrace bcc script
Add a script for recording per-lock statistics and write the result
to a csv file. See example file. The output is intended to be processed
by another script / tool.
One such script is lockstat.py, used to view the data in various
forms. For usage see examples file.

I've separated out the .c file so that it can be used in a C++ tool as
well.