Fix bpf_trace_printk() for big-endian targets
The format string for bpf_trace_printk() gets garbled during the
conversion of the LLVM IR code to the corresponding BPF object.
This happens because bcc explicitly sets the data layout of the
module to 'e-m:e-p:64:64-i64:64-n32:64-S128' which is valid for
little-endian targets only. If the arch is set to 'bpf', LLVM
selects the correct data layout automatically and there is no
need to set this explicitly.

Signed-off-by: Sandipan Das <sandipan@linux.vnet.ibm.com>
Reported-by: Chandan Rajendra <chandan@linux.vnet.ibm.com>
Fixes: #1723