Correction of an ipdb initdb issue
When using ns_ipdb.initdb(), the NetNS object should be given as a
parameter. In this case, the IPDB object is attached to the default ns.
Therefore, there is a key error exception in the next line :
in_ifc = ns_ipdb.interfaces[in_ifname]
however IPDB is updated asynchronously but the commit() operation is synchronous
So there is no risk to remove the initdb instruction